Finished

+++ queue.c
@@ -3,7 +3,7 @@
 #include "queue.h"
 
 // A utility function to create a new linked list node. 
-QNode* newNode(long int k) 
+QNode* newNode(int k) 
 { 
     QNode* temp = (QNode*)malloc(sizeof(QNode)); 
     temp->key = k; 
@@ -20,7 +20,7 @@ Queue* createQueue()
 } 
   
 // The function to add a key k to q 
-void enQueue(Queue* q, long int k) 
+void enQueue(Queue* q, int k) 
 { 
     // Create a new LL node 
     QNode* temp = newNode(k); 
@@ -37,7 +37,7 @@ void enQueue(Queue* q, long int k)
 } 
 
 // Function to remove a key from given queue q 
-long int deQueue(Queue* q) 
+int deQueue(Queue* q) 
 { 
     // If queue is empty, return NULL. 
     if (q->front == NULL)
